// apps/web/src/app/api/models/beziehung-antwort-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ImmobilieAntwortDto } from '../models/immobilie-antwort-dto';
import { KontaktAntwortDto } from '../models/kontakt-antwort-dto';
export interface BeziehungAntwortDto {
  beziehungstyp: number;
  dienstleistungstyp?: number;
  enddatum: string;
  id: string;
  immobilie: ImmobilieAntwortDto;
  kontakt: KontaktAntwortDto;
  startdatum: string;
}