// apps/web/src/app/api/models/immobilie-antwort-mit-beziehungen-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BeziehungenEingefuegtDto } from '../models/beziehungen-eingefuegt-dto';
export interface ImmobilieAntwortMitBeziehungenDto {
  beschreibung: string;
  beziehungen: Array<BeziehungenEingefuegtDto>;
  hausnummer: string;
  id: string;
  land: 'Deutschland' | 'Italien' | 'Frankreich';
  name: string;
  postleitzahl: string;
  stadt: string;
  strasse: string;
}