// apps/web/src/app/api/models/kontakt-antwort-mit-beziehungen-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BeziehungenEingefuegtDto } from '../models/beziehungen-eingefuegt-dto';
export interface KontaktAntwortMitBeziehungenDto {
  beziehungen: Array<BeziehungenEingefuegtDto>;
  hausnummer: string;
  id: string;
  land: 'Deutschland' | 'Italien' | 'Frankreich';
  name: string;
  postleitzahl: string;
  stadt: string;
  strasse: string;
}